Added functionality for homework 4
In particular, I added a configuration file. In this configuration file, the following can be specified: a directory to keep log files; a max number of old log files to keep; where the username file is kept; and whether the server allows PORT and PASV commands. Also added a sample config file in the programs/ directory at .ftpdlog